requirements.txt
# Core dependencies for backtester
pandas>=1.3.0,<3.0.0
numpy>=1.21.0,<2.0.0
matplotlib>=3.4.0,<4.0.0

# Technical analysis library
TA-Lib>=0.4.0

# Japanese font support for matplotlib
japanize-matplotlib>=1.1.0

# Data processing
scipy>=1.7.0,<2.0.0

# Optional: For enhanced visualization
seaborn>=0.11.0,<1.0.0

# Parameter optimization
optuna>=4.0.0
plotly>=6.0.0

# Configuration management
PyYAML>=6.0.0

# Database
pymongo>=4.0.0,<5.0.0

# Yahoo Finance data
yfinance>=0.2.0
curl-cffi>=0.5.0

# Web scraping for stock data
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0

# Dashboard support
streamlit>=1.20.0

# Environment management
python-dotenv>=0.19.0